requirements.txt
Flask
Flask-WTF
Flask-Script
Flask-SQLAlchemy
Flask-DebugToolbar
Flask-Migrate
PyMySQL
itsdangerous
permission
Pillow
six
raven
leancloud-sdk

# dev
livereload
nose
coverage
pylint
glob2

# deploy
gunicorn
Fabric
